Move Experience tests to proper test class
